IGNITE-13836 Presentation part moved to REST module from configuration module
