+++ modules/runner/pom.xml
@@ -61,7 +61,7 @@
         <dependency>
             <groupId>org.apache.ignite</groupId>
             <artifactId>ignite-rest</artifactId>
-            <version>3.0-SNAPSHOT</version>
+            <version>${project.version}</version>
             <scope>compile</scope>
         </dependency>
     </dependencies>
