+++ pom.xml
@@ -61,7 +61,7 @@
         <module>modules/configuration</module>
         <module>modules/configuration-annotation-processor</module>
         <module>modules/rest</module>
-        <module>modules/ignite-runner</module>
+        <module>modules/runner</module>
         <module>modules/cli-common</module>
         <module>modules/cli</module>
     </modules>
